requirements.txt
appnope==0.1.3
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
Flask==2.2.2
fonttools==4.36.0
image-quality==1.2.7
imageio==2.21.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
libsvm==3.23.0.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
pyzmq==23.2.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
stack-data==0.4.0
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
